requirements.txt
asgiref==3.5.2
attrs==22.1.0
bcrypt==4.0.1
black==22.10.0
click==8.1.3
Django==4.1.2
djangorestframework==3.14.0
exceptiongroup==1.0.0rc9
iniconfig==1.1.1
mypy-extensions==0.4.3
mysqlclient==2.1.1
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
pydantic==1.10.2
PyJWT==2.6.0
pyparsing==3.0.9
PyPika==0.48.9
pytest==7.2.0
pytest-django==4.5.2
pytz==2022.5
PyYAML==6.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
